add support for max number of runners
Now that we support multiple containers of the same handler, we should
also have the ability to limit the number of concurrent requests in a
container. This degenerates to serial reuse at 1 and 0 allows for
unlimited parallel reuse. This takes care of adding and removing from
the handler manager at appropriate times. Add when you are the first
of the maxRunners to finish and Remove when you are the last of the
concurrent runners to start.