Wait for thawing in olcontainer Unpause()
We were returning from Unpause() immediately, which doesn't
guarantee that the container is actually unpaused when we forward
the request. This change polls the freezer state file to wait
until it is marked as thawed by the kernel. This is slow, and we
should think of a way to improve it.