import-cache: do not hold locks when waiting for a cache entry to init
Prior, both the CacheManager lock and the ForkServer lock were held
while forking a new cache entry and waiting for it to come up. Now, we
release the locks, but must reacquire the CacheManager when adding it to
the CacheManager's server list and hold the lock so that while we
forkRequest, the evictor cannot kick us out.