Properly clean up after import cache SOCKs
We were only killing the init processes in SOCK containers, but
this does not clean up other processes when we do not use the PID
namespace (i.e., for containers in the import cache). Instead, we
explicitly SIGKILL all of the processes in the container.

Also made a util package with the goal of moving common code there.