Clean up recv_fds
Python's recv_fds function (https://github.com/python/cpython/blob/3.12/Lib/socket.py#L565) is almost the same as the one that was implemented here. The only notable difference is Python's function returning 4 variables instead of 2, so all that has to be done is modify the assignment to handle 4 variables being returned.