Let an idle LambdaInstance automatically pick up a kill signal
In the current implementation, as LambdaInstances are sent a kill signal,
their clean up is sent to a separate buffered cleanup channel. Therefore,
when LambdaFunction wants to scale down, it is imperitive that it actually
sends idle invocations to be cleaned up. Currently that is not the case, we
send the latest invocation to be cleaned up. The simplest solution is to
simply send one kill signal to a shared channel, and the idle invocation
will pick it up.